==> 16S/03_permafrostPS.qmd <==
---
title: "03.permafrostPS"
format: html
editor: visual
editor_options: 
  chunk_output_type: console
---

## Setup

```{r}
library(phyloseq) # Handling and analysis of high-throughput microbiome census data, Bioconductor v1.46.0
library(tidyverse) # Easily Install and Load the 'Tidyverse', CRAN v2.0.0
library(microViz) # Microbiome Data Analysis and Visualization, [::david-barnett/microViz] v0.12.1
library(patchwork)
```

## Load PS object

```{r}
ps <- readRDS("permafrost16S_processed_filt_ps.rds")

# Validate the PS object
phyloseq_validate(ps)
```

phyloseq checks that your sample and taxa names are consistent across the different slots of the phyloseq object. microViz provides [`phyloseq_validate()`](https://david-barnett.github.io/microViz/reference/phyloseq_validate.html) to check for and fix other possible problems with your phyloseq that might cause problems in later analyses. It is recommended to run this at the start of your analyses, and fix any problems identified.

```{r}
ps <- tax_fix(ps)

# Validate the PS object
phyloseq_validate(ps)
```

## Load PLFA abundance data

We can estimate the total biomass in a sample with our PLFA extraction data, which was quantified by GC-FID.

```{r}
PLFA_data <- read_rds("data/total_PLFA.rds") |> 
  rename(depth = core,
         incubation_temp_C = temp) |> 
  mutate(incubation_temp_C = as.factor(incubation_temp_C),
         depth = case_when(depth == "35m" ~ "35",
                           depth == "54m" ~ "54",
                           depth == "83m" ~ "83",
                           TRUE ~ depth),
         depth = as.factor(depth),
         )
```

## Separate blanks from samples

```{r}
ps.samples <- ps |> 
  ps_mutate(
    incubation_temp_C = factor(incubation_temp_C, levels = c(-4, 4, 12)),
    incubation_time_d = factor(incubation_time_d, levels = c(180, 30, 7, 0))
  ) |>  
  ps_filter(!is.na(depth)) |> 
  ps_mutate(inc_time_d = as.numeric(as.character(incubation_time_d))) |> 
  ps_mutate(inc_time_d_str = paste0(inc_time_d, " days")) |> 
  ps_mutate(inc_time_d_fct = factor(
    inc_time_d_str, 
    levels = c("0 days", "7 days", "30 days", "180 days"))) |> 
  ps_mutate(inc_temp_C_str = factor(paste0(incubation_temp_C, "C"), levels = c("-4C", "4C", "12C"))) |> 
  ps_arrange(inc_time_d_fct)

sample_names(ps.samples)

sample_data(ps.samples)
```

```{r}
ps.blanks <- ps |> 
  ps_filter(is.na(depth))

sample_names(ps.blanks)
```

## Inspect the blanks

```{r}
ps.blanks |> plot_bar(fill = "Phylum")
```

There appears to be negligible contamination in our negative controls: between 20 and 30 reads.

## Barplot

```{r}
p_barplot <- ps.samples |> 
  ps_mutate(
    inc_time_d_fct = factor(
      inc_time_d_fct, levels = c("180 days", "30 days", "7 days", "0 days"))
  ) |> 
  tax_agg("Genus") |> 
  comp_barplot(tax_level = "Genus",
               x = "inc_time_d_fct",
               n_taxa = 40,
               ) +
  #geom_area() +
  coord_flip() +
  facet_grid(vars(depth, incubation_temp_C)) +
  labs(x = "Incubation time (days)") +
  theme(
    legend.position = "left"
  ) +
  guides(
    fill = guide_legend(ncol = 2)
  )
p_barplot
```

## Archaea Barplot

```{r}
interestingGenera <- ps.samples |> tax_select("Archaea") |> tax_top(n = 10, rank = "Genus")

p_archaea_barplot <- ps.samples |> 
  ps_mutate(
    inc_time_d_fct = factor(
      inc_time_d_fct, levels = c("180 days", "30 days", "7 days", "0 days"))
  ) |> 
  comp_barplot(
  tax_level = "Genus",
  n_taxa = 7,
  palette = distinct_palette(7, pal = "kelly", add = "grey90"),
  tax_order = interestingGenera,
  x = "inc_time_d_fct",
) +
  coord_flip(ylim = c(0, 0.045)) +
  facet_grid(vars(depth, incubation_temp_C)) +
  labs(
    x = "Incubation time (days)"
  ) +
  theme(
    legend.position = "left"
  )
p_archaea_barplot

cowplot::save_plot(
  filename = "output/p_archaea_barplot.pdf",
  plot = p_archaea_barplot,
  base_height = 6,
  base_width = 9
)
  
```

## Alpha Diversity (Simpson's)

Alpha diversity describes "within-sample" diversity. It's a measure of how diverse a single sample is, not taking into account observations of other samples.

Here we apply the Simpson index (or the "Gini-Simpson index"). The Simpson index is similar to the [Shannon Index](https://en.wikipedia.org/wiki/Diversity_index#Shannon_index): it is based on the probability that two entities (microbes, reads, etc.) taken from the sample at random are of different types (e.g., species). As this is a probability, the score ranges from 0 to 1.

$$
Simpson = 1 - \sum_{i = 1}^{R} p_i^2
$$

where $p_i$ is the proportion of observations (reads) belonging to the $i$th species.

```{r}
p_alpha <- ps.samples |>
  ps_arrange(inc_time_d_fct) |> 
  plot_richness(
    x = "inc_time_d_fct",
    measures = "Simpson",
    color = "depth"
  ) +
  facet_wrap(vars(incubation_temp_C)) +
  geom_line(aes(group = depth)) +
  labs(
    x = "Incubation Time (Days)",
    y = "Simpson Index"
  ) +
  theme_bw() +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1)
    )
p_alpha

p_alpha2 <- ps.samples |>
  ps_mutate(
    inc_time_d_fct = factor(inc_time_d_fct, levels = c("180 days", "30 days", "7 days", "0 days"))
  ) |> 
  plot_richness(
    x = "inc_time_d_fct",
    measures = "Simpson",
  ) +
  geom_line(aes(group = depth), color = "black") +
  geom_point(aes(fill = incubation_temp_C, shape = depth), color = "black") +
  facet_grid(vars(depth, incubation_temp_C)) +
  coord_flip() +
  scale_color_viridis_d() +
  scale_fill_viridis_d() +
  scale_shape_manual(values = c(21, 22, 23, 24)) +
  labs(
    x = "",
    y = "Simpson Index"
  ) +
  theme_bw() +
  theme(
    strip.background = element_rect(fill = "black"),
    strip.text = element_text(color = "white", face = "bold", size = 2),
    legend.position = "None",
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank()
    )
p_alpha2



cowplot::save_plot(
  filename = "output/p_alpha.pdf",
  plot = p_alpha,
  base_height = 5,
  base_width = 5
)
```

### Combined Alpha and barplot

```{r}
p_barplot_nostrip <- p_barplot %+% theme(strip.text = element_blank(), strip.background = element_blank())

p_combined <- p_barplot_nostrip + p_alpha2 + plot_layout(widths = c(5, 1))

# cowplot::save_plot(
#   plot = p_combined,
#   filename = "output/p_combined_bar_alpha.pdf",
#   base_height = 10,
#   base_width = 16
# )
```

# Ordination

Bray-Curtis dissimilarity is a statistic used to quantify the dissimilarity in species composition between sites based on species counts. The Bray-Curtis dissimilarity $BC_{jk}$ between two sites, noted as j and k, is defined as:

$$
BC_{jk} = 1 - \frac{2C_{jk}}{S_j + S_k} = 1 - \frac{2 \sum_{i=1}^{p}min(N_{ij}, N_{ik})}{\sum_{i = 1}^{p} (N_{ij} + N_{ik})}
$$

Where $N_{ij}$ is the number of specimens i at site j, $N_{ik}$ is the number of specimens i at site k, and p is the total number of species in all the samples.

BC dissimilarity is bounded between zero and one: 0 means the two sites have the same composition (sharing all species). 1 means that the two sites do not share any species.

BC dissimilarity is sometimes incorrectly referred to a distance, but is in fact a dissimilarity. A distance obeys triangle inequality.

```{r}
p_bray <- ps.samples |> 
  ps_arrange(inc_time_d_fct) |> 
  tax_transform("identity", rank = "Genus") |> 
  dist_calc("bray") |> 
  ord_calc("PCoA") |> 
  ord_plot() +
  ggforce::geom_link2(
    aes(color = inc_time_d_fct, group = depth),
    linewidth = 0.8, n = 500
  ) +
  geom_point(aes(fill = inc_time_d_fct, shape = depth), 
             size = 3, color = "black", stroke = 1) +
  scale_shape_manual(values = c(21, 22, 23, 24)) +
  scale_color_viridis_d(option = "magma", end = 0.9) +
  scale_fill_viridis_d(option = "magma", end = 0.9) +
  facet_wrap(vars(inc_temp_C_str)) +
  labs(
    color = "Incubation time",
    shape = "Site"
  ) +
  guides(
    fill = "none"
  ) +
  theme_bw() +
  theme(
    aspect.ratio = 1,
    strip.background = element_blank(),
    strip.text = element_text(color = "black", face = "bold", size = 14)
  )
p_bray

cowplot::save_plot(
  filename = "output/p_bray.png",
  plot = p_bray,
  base_height = 5,
  base_width = 11
)
```

# Calculate biomass-normalized abundance

This is an estimation of taxonomic absolute abundance which multiplies relative abundance by the total amount of intact biomass in the sample, as a first order approximation.

$$
x\: pg PLFA \times \frac{1 \: cell}{0.003503333 \: pg \: PLFA} = y \: cells
$$

```{r}
# PLFA to cell conversion factor: 0.003503333 pg PLFA per cell
# See Frostegaard and Baath reference
PLFA_data <- PLFA_data |> 
  mutate(
    total_PLFA_pg_per_g = total_PLFA_ng_per_g * 1000, # 1000 pg / ng
    cells_per_g = total_PLFA_pg_per_g * (1/0.003503333)  # x PLFA * (cell/PLFA) = ycell
  )


ps.samples.PLFA_melted <- ps.samples |> 
  ps_melt() |> 
  left_join(PLFA_data, by = c("depth", "incubation_temp_C", "inc_time_d")) |> 
  group_by(Sample) |> 
  mutate(
    total_reads = sum(Abundance, na.rm = TRUE),
    relative_abundance = Abundance / total_reads,
    biomass_normalized_abundance = relative_abundance * total_PLFA_ng_per_g
  )
```

```{r}
# Create dataset of relative abundance scaled by PLFA
ps.samples.PLFA <- ps.samples |> 
  ps_join(PLFA_data, by = c("depth", "incubation_temp_C", "inc_time_d")) |> 
  # transform to relative abundance
  transform_sample_counts(function(x){x/sum(x)})

for(n in 1:nsamples(ps.samples.PLFA)) {
  otu_table(ps.samples.PLFA)[,n] <- otu_table(ps.samples.PLFA)[,n]*sample_data(ps.samples.PLFA)$total_PLFA_ng_per_g [n]
}

# Create dataset of relative abundance scaled by estimated cell density
ps.samples.cells <- ps.samples |> 
  ps_join(PLFA_data, by = c("depth", "incubation_temp_C", "inc_time_d")) |> 
  # transform to relative abundance
  transform_sample_counts(function(x){x/sum(x)})

for(n in 1:nsamples(ps.samples.cells)) {
  otu_table(ps.samples.cells)[,n] <- otu_table(ps.samples.cells)[,n]*sample_data(ps.samples.cells)$cells_per_g [n]
}

# Plot PLFA weighted abundance
p_ps.samples_PLFA <- ps.samples.PLFA |> 
  ps_mutate(
    inc_time_d_fct = factor(
      inc_time_d_fct, levels = c("180 days", "30 days", "7 days", "0 days"))
  ) |> 
  tax_agg("Genus") |> 
  comp_barplot(tax_level = "Genus",
               tax_transform_for_plot = "identity",
               x = "inc_time_d_fct",
               n_taxa = 40,
               ) +
  coord_flip() +
  facet_grid(vars(depth, incubation_temp_C), scales = "free_x") +
  labs(x = "Incubation time (days)",
       y = latex2exp::TeX("Biomass-Weighted Abundance (RA \\times ng/g)")
       ) +
  theme(
    legend.position = "left"
  ) +
  guides(
    fill = guide_legend(ncol = 2)
  )
p_ps.samples_PLFA

p_PLFA <- PLFA_data |> 
  mutate(inc_time_d_str = paste0(inc_time_d, " days")) |> 
  mutate(inc_time_d_fct = factor(inc_time_d_str, levels = c("180 days", "30 days", "7 days", "0 days"))) |> 
  ggplot() +
  aes(
    x = inc_time_d_fct,
    y = total_PLFA_ng_per_g / 1000,
    fill = as.factor(incubation_temp_C)
  ) +
  geom_path(aes(group = depth)) +
  geom_point(shape = 21) +
  scale_fill_viridis_d() +
  scale_y_continuous() +
  coord_flip() +
  facet_grid(vars(depth, incubation_temp_C)) +
  labs(y = "PLFA Biomass (µg/g)", x = "") +
  theme_bw() +
  theme(
    legend.position = "None",
    strip.text = element_text(size = 2),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank()
  )
p_PLFA
```

# New combined plot

```{r}
p_combined_PLFA <- cowplot::plot_grid(
  p_ps.samples_PLFA, p_alpha2,
  nrow = 1, rel_widths = c(0.85, 0.15)
)
p_combined_PLFA

cowplot::save_plot(
  filename = "output/combined_barplot_w_PLFA.pdf",
  plot = p_combined_PLFA, base_height = 10, base_width = 14
)
```
